Fixed a bug for audio settings on game start
Audio settings were not being loaded appropriately at the beginning of the game. So, I added some logic on the character to do so in their Begin Play.

In the event audio settings have already been loaded and we call the load audio function on the game instance, the game instance will now exit early without trying to load the save game.