Desert Level Updates
Adjusted lighting in the gravity maze planet. Added torches to indicate the direction gravity is flowing, to help the player progress through the planet. Added visual indicators to the labyrinth to help guide the player in the correct direction. Moved the sand block in the labyrinth to line up with the labyrinth and eliminate visual gaps.